requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.1.0
certifi==2025.1.31
coverage==7.6.12
h11==0.14.0
idna==3.10
iniconfig==2.0.0
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
